requirements.txt
Flask==2.3.5
Authlib==1.2.1
SQLAlchemy==2.0.23
python-dotenv==1.0.0
gunicorn==21.2.0